# backend-api/swagger/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Alumni CW API",
    "version": "1.0.0",
    "description": "Backend-only CodeIgniter 3 API for alumni authentication, profile management, blind bidding, featured alumnus selection, and API key protected public access."
  },
  "servers": [
    {
      "url": "http://localhost/alumni-cw/index.php",
      "description": "Local development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      }
    },
...
  },
  "paths": {
...
    "/api/auth/verify-email": {
      "get": {
        "tags": ["Auth"],
        "summary": "Verify email using token",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "example": "verification_token_here"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Email verified successfully"
          },
          "400": {
            "description": "Invalid or expired token"
          }
        }
      }
    },
...
  }
}